# packages/poster-contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Social Poster internal API",
    "version": "1.4",
    "description": "Cross-platform publishing API. Two auth modes share every /v1 route (D-023): app mode uses a client-credentials JWT, user mode uses the signed-in user’s Supabase session JWT and acts as the first-party poster-web app."
  },
  "servers": [
    {
      "url": "/",
      "description": "The service root; deployments front this with TLS."
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "App mode: the token from POST /v1/oauth/token. User mode: a Supabase session JWT."
      }
    },
    "schemas": {
...
      "Media": {
        "type": "object",
        "properties": {
          "media_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["pending_upload", "ready", "failed"],
            "description": "pending_upload until the bytes arrive and are probed."
          },
          "kind": {
            "type": "string",
            "enum": ["image", "video"]
          },
          "mime_type": {
            "type": "string"
          },
          "size_bytes": {
            "nullable": true,
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "duration_s": {
            "nullable": true,
            "type": "number"
          },
          "width": {
            "nullable": true,
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 9007199254740991
          },
          "height": {
            "nullable": true,
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 9007199254740991
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "media_id",
          "status",
          "kind",
          "mime_type",
          "size_bytes",
          "duration_s",
          "width",
          "height",
          "created_at"
        ],
        "additionalProperties": false,
        "description": "A stored media item."
      },
      "SignedUploadRequest": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "kind": {
            "type": "string",
            "enum": ["image", "video"]
          },
          "mime_type": {
            "type": "string",
            "minLength": 1
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 9007199254740991,
            "description": "Declared size, used to reject uploads the plan does not allow."
          },
          "filename": {
            "description": "Only used to pick a storage extension.",
            "type": "string"
          }
        },
        "required": ["user_id", "kind", "mime_type", "size_bytes"],
        "additionalProperties": false,
        "description": "Ask for a signed URL instead of sending the bytes through the API."
      },
      "SignedUploadResponse": {
        "type": "object",
        "properties": {
          "media_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["pending_upload"]
          },
          "upload_url": {
            "type": "string",
            "format": "uri",
            "description": "PUT the bytes here, then call /complete."
          },
          "expires_at": {
            "type": "string"
          }
        },
        "required": ["media_id", "status", "upload_url", "expires_at"],
        "additionalProperties": false,
        "description": "A one-shot upload target. Nothing is usable until /complete succeeds."
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "invalid_request",
                  "invalid_token",
                  "forbidden_user",
                  "grant_missing",
                  "not_found",
                  "idempotency_conflict",
                  "too_late",
                  "constraint_violation",
                  "rate_limited",
                  "internal_error"
                ],
                "description": "Machine-readable error code."
              },
              "message": {
                "type": "string",
                "description": "Human-readable summary. Never used for control flow."
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "target_index": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991,
                      "description": "Index of the failing target in the submitted targets array."
                    },
                    "connection_id": {
                      "description": "Public id of the target connection.",
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "description": "Constraint code; see the contract §8 list."
                    },
                    "constraint": {
                      "description": "The limit that was exceeded and the actual value, as data.",
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": ["target_index", "code"],
                  "additionalProperties": false,
                  "description": "One per failing target, so a client can fix exactly what broke."
                }
              },
              "request_id": {
                "type": "string",
                "description": "Echoes the X-Request-Id response header. Quote it in support requests."
              }
            },
            "required": ["code", "message", "request_id"],
            "additionalProperties": false
          }
        },
        "required": ["error"],
        "additionalProperties": false,
        "description": "The single error envelope used by every non-2xx response."
      }
    }
  },
  "paths": {
...
    "/v1/media": {
      "post": {
        "operationId": "createMedia",
        "summary": "Upload media, or get a signed URL to upload it directly",
        "description": "Two paths, chosen by Content-Type. `multipart/form-data` sends the bytes through the API, which stores and probes them and returns a ready media row; it is capped in size. `application/json` returns a signed upload URL instead, for large video that should not pass through the API process — PUT the bytes to that URL and then call /complete. Media is uploaded once and referenced by id from any number of posts (FR-07).",
        "tags": ["media"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["user_id", "file"],
                "properties": {
                  "user_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignedUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stored and probed (multipart path).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            }
          },
          "202": {
            "description": "Signed URL issued; nothing is usable until /complete.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request, unsupported type, or file too large.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or expired token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "User-mode token does not match the requested user_id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Per-app rate limit exceeded; honor Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/media/{media_id}/complete": {
      "post": {
        "operationId": "completeMedia",
        "summary": "Finish a signed-URL upload",
        "description": "Probes the uploaded object for duration and dimensions and marks the media ready. Until this succeeds the media cannot be posted, and referencing it yields the `media_not_ready` constraint code.",
        "tags": ["media"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Probed and ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            }
          },
          "400": {
            "description": "The object is missing or could not be probed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed, or expired token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such media for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}